Assignment that handles data about countries.
It answers the following questions :
How many countries in ASIA?
How many countries in EUROPE?
How many countries with a Literacy Rate > 70
How many countries with a GDP > 30000
How many countries with a GDP > 20000 and a literacy rate < 90
Which is the most populous country in the world?
Which is the least populous country in the world?
Which is the largest country in the world?(by area)
Which is the smallest country in the world?(by area)
What is the total area of all the countries of the world?
Which are all the landlocked countries of the world?